correct link pollution bug spawned from reusing map-integrated event summary component
